Add RePlay! logo through play.plugins file
It was added to the original Play! 1 location: "app/play.plugins",
but it doesn't get processed by the maion java sourceset.

This "processResources" is just a workaround to add "play.plugins" to
the root of the classpath while maintaining the old directory structure.

$ find build/ -name play.plugins
build/resources/main/play.plugins

More to read at: "gradle - How Do I Add Multiple Directories And Filter
Patterns To A SourceSet?" [1]

[1]: https://stackoverflow.com/q/63626566